Refactoring and fix circular imports
