+++ argopy/__init__.py
@@ -21,29 +21,27 @@ import logging
 log = logging.getLogger(__name__)
 log.addHandler(logging.NullHandler())
 
-# Import facades:
-from .fetchers import ArgoDataFetcher as DataFetcher  # noqa: E402
-from .fetchers import ArgoIndexFetcher as IndexFetcher  # noqa: E402
-
-from .xarray import ArgoAccessor  # noqa: E402
-
-# Other Import
-# from . import utils as utilities  # noqa: E402
-from . import utilities  # noqa: E402  # being deprecated until 0.1.15, then remove
+# Low-level imports
 from . import stores  # noqa: E402
 from . import errors  # noqa: E402
 from . import plot  # noqa: E402
 from . import tutorial  # noqa: E402
 from .plot import dashboard, ArgoColors  # noqa: E402
 from .options import set_options, reset_options  # noqa: E402
-from .data_fetchers import CTDRefDataFetcher  # noqa: E402
 from .stores import ArgoIndex  # noqa: E402
 from .utils import show_versions, show_options  # noqa: E402
 from .utils import clear_cache, lscache  # noqa: E402
 from .utils import MonitoredThreadPoolExecutor  # noqa: E402, F401
 from .utils import monitor_status as status  # noqa: E402
-from .related import TopoFetcher, OceanOPSDeployments, ArgoNVSReferenceTables, ArgoDocs, ArgoDOI  # noqa: E402
 
+# High-level imports:
+from .related import TopoFetcher, OceanOPSDeployments, ArgoNVSReferenceTables, ArgoDocs, ArgoDOI, CTDRefDataFetcher  # noqa: E402
+from .fetchers import ArgoDataFetcher as DataFetcher  # noqa: E402
+from .fetchers import ArgoIndexFetcher as IndexFetcher  # noqa: E402
+
+from .xarray import ArgoAccessor  # noqa: E402
+
+from . import utilities  # noqa: E402  # being deprecated until 0.1.15, then remove
 
 #
 __all__ = (
